Every structure declares its own vocabulary; dissolve `buildin`
Two changes that turn out to be one. `.symbolic()` could be called with no
arguments on nine of the library's structures, because each declares a
`default_signature` naming its operators and the codec that turns Python values
into elements. It could not on Büchi arithmetic or MSO0: those were built by
plain functions returning a bare presentation, with nowhere to hang a
signature. `autstr.arithmetic` existed to fill that gap -- for ONE of the three,
which is what made it look like a second, rival interface.

So the signature moves onto the presentation, where the rest of the library
keeps it, and the module that worked around its absence goes away. All three
gain their codec, verified against the automata rather than assumed:

  - N: binary, least significant bit first (checked against A(x,y,z)),
  - Z: sign symbol then magnitude (as before),
  - MSO0: a set as a bitmask, position i set iff i is a member, in CANONICAL
    form -- the universe rejects trailing zeros, so {0} is `1` and the empty
    set is the empty word (checked against Subset over all subsets of {0,1,2}).

MSO0 therefore gains an interface it never had: union, intersection and
difference as `+`, `*`, `-`, and solutions that come back as Python sets.

`buildin` was the other half of the same confusion: every structure the library
ships is built in, so the name carved nothing. Its contents move to where their
subject already lives -- `arithmetic` (Buechi N and Z), `powerset` (MSO0),
`tree_arithmetic` (Skolem, matching algebra/tree_algebra), and the generic
automaton constructors to `utils.automata_tools`, which was already importing
`one` from across the package boundary. `CompiledPresentation` holds the part
the presentations share.

The precompiled `.autstr` artifacts go too. They saved 0.03s, 0.23s and 0.44s
against building from scratch, and the built presentations are identical to the
loaded ones relation by relation. Serialization stays fully supported; it just
stops shipping artifacts nobody needed.

BREAKING: `autstr.buildin.*` is gone, as are `autstr.arithmetic.integers` and
`signature`; encode/decode are now classmethods on the presentations.
