+++ tests/test_interpretations.py
@@ -3,8 +3,10 @@ import itertools
 
 import pytest
 
-from autstr.arithmetic import decode, encode
-from autstr.buildin.presentations import BuechiArithmeticZ
+from autstr.arithmetic import BuechiArithmeticZ
+
+encode, decode = BuechiArithmeticZ.encode, BuechiArithmeticZ.decode
+from autstr.arithmetic import BuechiArithmeticZ
 from autstr.infinite_graphs import IntegerGrid, InfiniteGraph
 from autstr.interpretations import interpret
 from autstr.symbolic import FunctionCodec
